Added random sampling antialiasing.
Also, made the pixels get clamped to a maximum brightness value.
This is part of the random sampling, kind of, since now there's a step
where a subpixel value can be greater than 1.0, and I guess it's to
prevent floating point overflow (which I think goes to infinity?) from
causing the pixels to be invalid. Or, it's to guarantee the same thing
if the number of samples is incorrectly supplied.
Either way, seems like it just saves us from issues we wouldn't be able
to recover from anyway.