amend! Metal shader!
Metal shader!

Not quite right, and also I don't fully understand the reflect function,
but I wrote this scatter function myself (that's the bit that's not
quite right)! Also, I discovered that cargo run --release is very
definitely worth it here (it gives a 32x speedup, at least on bigimage,
which is 1080p).