Made the logic work.
It's a bit ugly, though, but I can't think of how to improve it without
nesting things or making things mutable which shouldn't be.
Thankfully I think the compiler can probably reason about what needs to
be computed when, but it's still a bit hard to read, especially with an
early return in a match statement.