Used rustfmt on a bit of code that bothered me.
Unfortunately, what it did is almost as bad as what was being done
before... but in a very different way.
More unfortunately, it *really* doesn't like to look only at parts of
files, so using it in nvim to format only part of a file is not really
generally feasible.
That's just a shame. I am proud of using it in nvim to "filter" the
current selection, though. I must've seen that pattern somewhere...