Factored out basic scene generator into own binary.
I tried to do this sorta thing with jaq, but it was just too hard, and
this is more flexible anyway. I'm not sure the Cargo workspace is set up
quite right, but it's a start. And hey, I guess my raytracer is a
library now!

This was motivated primarily to let me randomly generate emissive
scenes, so that they might be better lit than the current randomly
generated one.