Used multiple threads, made some timings.
Timings are counter to my expectations - the uneven, normal images are
faster threaded than with rayon, while the more balanced crazy images
are faster with rayon. Maybe just because the normal images are so much
quicker for both, so overhead matters more their? Does rayon get better
for longer runtimes?