[SEMI UGLY] Added a first attempt at plane intersection.
The planes appear to show up fine. The problem is the
lighting/shading/materials - none of it works. My first guess is that
the problem is with normals, but I can't be bothered + don't have time
to fix it.

I just got way too excited and ported this from my javascript raytracer
that I just started on for Graphics. Now I need to go home and sleep.